优化目录结构

+++ lib/public_functions.py
@@ -1,6 +1,6 @@
 # -*- coding:utf-8 -*-
 
-import os,re,time
+import os, re, time
 import subprocess
 from pathlib import Path
 import yaml
@@ -12,7 +12,7 @@ from email.mime.text import MIMEText
 from email.mime.multipart import MIMEMultipart
 
 
-def send_mail(receivers,message):
+def send_mail(receivers, message):
     mail_host = "mail.sjdd.com.cn"
     mail_user = 'liml@sjdd.com.cn'
     mail_pass = 'sjdd1234'
@@ -20,7 +20,7 @@ def send_mail(receivers,message):
     sender = 'liml@sjdd.com.cn'
 
     parsreceivers = [f'{name}@sjdd.com.cn' for name in receivers]
-    with open(message,'rb') as f:
+    with open(message, 'rb') as f:
         body = f.read()
 
     message = MIMEMultipart()
@@ -43,20 +43,20 @@ def send_mail(receivers,message):
         print('error:', e.args)
 
 
-def get_real_dir_path(path,paths=''):
+def get_real_dir_path(path, paths=''):
     curdir = os.path.dirname(path)
     if paths is None:
         return curdir
     else:
-        return os.path.abspath(os.path.join(curdir,paths))
+        return os.path.abspath(os.path.join(curdir, paths))
 
 
 def 获取当前系统():
     systeminfo = platform.platform()
     # print(systeminfo)
-    if 'Windows' in systeminfo :
+    if 'Windows' in systeminfo:
         systemname = 'windows'
-    elif 'Darwin' in systeminfo :
+    elif 'Darwin' in systeminfo:
         systemname = 'mac'
     else:
         systemname = 'linux'
@@ -65,20 +65,21 @@ def 获取当前系统():
 
 def 获取控件文件信息(filename='app控件'):
     控件文件路径 = Path(__file__).cwd().parent / f'{filename}.yml'
-    with open(控件文件路径,'r', encoding='gbk') as loadfile:
+    with open(控件文件路径, 'r', encoding='gbk') as loadfile:
         info = yaml.load(loadfile)
     return info
 
 
-def 启动appium_server(udid,port=4723):
+def 启动appium_server(udid, port=4723):
     if 'mac' in 获取当前系统():
         excute_cmd_base = "node /Applications/Appium.app/Contents/Resources/app/node_modules/appium/build/lib/main.js"
     else:
         excute_cmd_base = "appium"
-    excute_cmd = "{} -a 127.0.0.1 -p {} -U {} --session-override".format(excute_cmd_base,port,udid)
+    excute_cmd = "{} -a 127.0.0.1 -p {} -U {} --session-override".format(excute_cmd_base, port, udid)
 
     subprocess.Popen()
     print(excute_cmd)
+
 
 def cleanNodeProcess():
     是否mac = 'mac' in 获取当前系统()
@@ -97,18 +98,23 @@ def cleanNodeProcess():
         execute_cmd = 'tasklist|findstr node'
         cmd_res = subprocess.getoutput(execute_cmd)
         for res in cmd_res.split('\n'):
-            while 'node.exe' in res:
+            if 'node.exe' in res:
                 pid = re.split('\s+', res)[1]
                 # print(pid)
                 kill_cmd = f'taskkill /F /PID {pid}'
                 kill_res = subprocess.getoutput(kill_cmd)
                 print(kill_res)
 
-def waittimeout(element,timeout=10):
+
+def waittimeout(element, timeout=10):
     begintime = time.time()
     while begintime + timeout > time.time():
-        if element is not None:
+        是否停止 = element is not None
+        if 是否停止:
             break
+        else:
+            time.sleep(1)
+
 
 def getLocalTime():
     loctime = time.strftime('%Y-%m-%d-%H:%M:%S', time.localtime())
@@ -116,7 +122,6 @@ def getLocalTime():
 
 
 def get_android_app_info(app="iRoom"):
-
     # get device name
     deviceId = subprocess.getoutput('adb devices').split('\n')[1].split("\t")[0]
 
@@ -156,7 +161,7 @@ def get_ios_app_info():
     device_os_version = os.popen('ideviceinfo -k ProductVersion').readlines()
 
     desired_caps = {
-        'platformName':'IOS',
+        'platformName': 'IOS',
         'platformVesion': device_os_version[0].rstrip(),
         'deviceName': 'IPhone',
         'automationName': 'XCUITest',
@@ -171,41 +176,42 @@ def get_ios_app_info():
 
 def 获取控件信息(devicedriverinfo):
     控件文件 = 获取控件文件信息()
-    isIOS = 'desired' not in devicedriverinfo#判断运行的设备是否为IOS
+    isIOS = 'desired' not in devicedriverinfo  # 判断运行的设备是否为IOS
     if isIOS:
         控件信息 = 控件文件['iRoom_{}'.format(devicedriverinfo['platformName'])]
         deviceid = devicedriverinfo['udid']
     else:
         控件信息 = 控件文件['iRoom_{}'.format(devicedriverinfo['desired']['platformName'])]
         deviceid = devicedriverinfo['desired']['deviceName']
-    return 控件信息,deviceid
+    return 控件信息, deviceid
 
 
 class StartDriver():
 
-    def __init__(self,devicelist):
+    def __init__(self, devicelist):
         设备信息 = 获取控件文件信息('devices')
         self.aport = list(range(4723, 4800, 2))
         self.bport = list(range(4724, 4800, 2))
-        self.iosport = list(range(8100,8200,2))
+        self.iosport = list(range(8100, 8200, 2))
         self.devicelist = devicelist
         self.realdevice = [设备信息[device] for device in devicelist]
 
-
-    def startAppiumServer(self,i):
+    def startAppiumServer(self, i):
         appium_env = os.environ['APPIUM']
         是否mac系统 = 'mac' in 获取当前系统()
         excute_cmd_base = f"node {appium_env}/Resources/app/node_modules/appium/build/lib/main.js -a 127.0.0.1"
         # print(excute_cmd_base)
         uidkey = 'udid' if 'IOS' in self.realdevice[i]['platformName'] else 'deviceName'
 
-        deviceport = f'--webdriveragent-port {self.iosport[i]}' if 'IOS' in self.realdevice[i]['platformName'] else f'-bp {self.bport[i]}'
+        deviceport = f'--webdriveragent-port {self.iosport[i]}' if 'IOS' in self.realdevice[i][
+            'platformName'] else f'-bp {self.bport[i]}'
         # print(uidkey)
         excute_cmd = f"{excute_cmd_base} -p {self.aport[i]} {deviceport} -U {self.realdevice[i][uidkey]} --local-timezone --log-timestamp --command-timeout 3000"
 
         appiumlogpath = '/Users/liminglei/Desktop/appium/' if 是否mac系统 else 'c:/'
 
-        subprocess.Popen(excute_cmd,shell=True,stdout=open(f"{appiumlogpath}appiumlog_{self.realdevice[i][uidkey]}.txt",'w+'))
+        subprocess.Popen(excute_cmd, shell=True,
+                         stdout=open(f"{appiumlogpath}appiumlog_{self.realdevice[i][uidkey]}.txt", 'w+'))
 
         time.sleep(5)
 
@@ -226,17 +232,3 @@ class StartDriver():
 
         # print(pidlist)
         return pidlist
-
-
-def startMultAppiumServer(sd):
-
-    proc_list = []
-
-    for i in range(len(sd.devicelist)):
-        proc_list.append(multiprocessing.Process(target=sd.startAppiumServer,args=(i,)))
-
-    for pro in proc_list:
-        pro.start()
-
-    for pro in proc_list:
-        pro.join()
